Refactor sticker service to external file
